api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy>=2.0.34
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1
httpx>=0.27.2
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
google-api-python-client>=2.146.0
itsdangerous>=2.2.0
starlette>=0.38.6
python-multipart>=0.0.12
